Fixing login issue
Problem was caused (at least for me) by the use of secure cookies in the app. Because the app runs locally on HTTP, the cookies weren't being passed correctly. From the [express-session GitHub page](https://github.com/expressjs/session#cookiesecure):

> Note be careful when setting this to true, as compliant clients will not send the cookie back to the server in the future if the browser does not have an HTTPS connection.

Passport.js was failing to validate the return call to the /token endpoint because the state didn't match. This caused it to invoke the failureRedirect back to the homepage.

Also updated passport and passport-azure-ad to current versions and refactored the way users are handled to match.